requirements.txt
pandas==2.1.4
python-docx==0.8.11
google-generativeai==0.3.2
python-dotenv==1.0.0
undetected-chromedriver==3.5.4
selenium==4.15.2
requests==2.31.0
Pillow==10.1.0
xlsxwriter==3.1.9
openpyxl==3.1.2